image_app_fast/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
Pillow
requests
